+++ js/buildings/energySeller.js
@@ -1,7 +1,4 @@
-﻿var energySellerCost = 10;
-var energySellerCount = 0;
-
-class Seller {
+﻿class Seller {
     place(x, y) {
         var newEnergySeller = new Image();
         newEnergySeller.onload = function () {
@@ -36,5 +33,13 @@ class Seller {
             right: canvasWidth - 155 + 25 + canvasRect.left,
         };
         return location;
+    }
+
+    clicked(x, y) {
+        var iconLoc = this.icon();
+        if (x >= iconLoc.left && y >= iconLoc.top && x <= iconLoc.right && y <= iconLoc.bottom) {
+            return true;
+        }
+        return false;
     }
 }