+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v1.0.2
+- Added energy and seller upgrades.
+- Gain rates will now display in the information box.
 ## v1.0.1
 - Tidied up code a bit.
 - Fixed reset button not resetting game.